+++ src/NpgsqlAnalyzers/NpgsqlAnalyzer.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Data;
+using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using Microsoft.CodeAnalysis;
@@ -14,33 +16,59 @@ namespace NpgsqlAnalyzers
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public class NpgsqlAnalyzer : DiagnosticAnalyzer
     {
-        private readonly string _connectionString;
+        private const string ConfigFileName = ".npgsqlanalyzers";
+
+        private Configuration _configuration;
 
         public NpgsqlAnalyzer()
-            : this(Configuration.ConnectionString)
         {
         }
 
-        public NpgsqlAnalyzer(string connectionString)
+        public NpgsqlAnalyzer(Configuration configuration)
         {
-            _connectionString = string.IsNullOrWhiteSpace(connectionString)
-                ? throw new InvalidOperationException("Invalid connection string.")
-                : connectionString;
+            _configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
         }
 
         public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } = ImmutableArray.Create(
             Rules.BadSqlStatement,
             Rules.UndefinedTable,
-            Rules.UndefinedColumn);
+            Rules.UndefinedColumn,
+            Rules.MissingCommand);
 
         public override void Initialize(AnalysisContext context)
         {
             context.EnableConcurrentExecution();
             context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
 
-            context.RegisterSyntaxNodeAction(
-                AnalyzeInvocationExpressionNode,
-                SyntaxKind.ObjectCreationExpression);
+            context.RegisterCompilationStartAction((compilationContext) =>
+            {
+                if (_configuration is null)
+                {
+                    var configFile = compilationContext.Options.AdditionalFiles.FirstOrDefault(
+                        (file) => Path.GetFileName(file.Path).Equals(ConfigFileName));
+                    if (configFile is null)
+                    {
+                        throw new InvalidOperationException("Missing configuration file.");
+                    }
+
+                    _configuration = Configuration.FromFile(
+                        configFile.GetText().Lines.Select(line => line.ToString()));
+                }
+
+                compilationContext.RegisterSyntaxNodeAction(
+                    AnalyzeInvocationExpressionNode,
+                    SyntaxKind.ObjectCreationExpression);
+            });
+        }
+
+        private static void Log(string value)
+        {
+            var logFile = Path.Combine(
+                Environment.GetFolderPath(Environment.SpecialFolder.Desktop),
+                nameof(NpgsqlAnalyzer),
+                "log.txt");
+            Directory.CreateDirectory(Path.GetDirectoryName(logFile));
+            File.AppendAllText(logFile, $"{value}{Environment.NewLine}");
         }
 
         /// <summary>
@@ -55,7 +83,7 @@ namespace NpgsqlAnalyzers
         /// <returns>
         /// The pure query, without the enclosing quotes and @.
         /// </returns>
-        private static string SanitizeQuery(string queryLiteral) =>
+        private static string ExtractQuery(string queryLiteral) =>
             queryLiteral
                 .Trim()
                 .Substring(1) // Removes the @ or " at the start of the string definition
@@ -73,8 +101,163 @@ namespace NpgsqlAnalyzers
         private static string ReplaceNamedParameters(string query) =>
             Regex.Replace(query, @"@\w+", "NULL");
 
-        private static string ExtractQuery(string literal) =>
-            ReplaceNamedParameters(SanitizeQuery(literal));
+        private static VariableDeclaratorSyntax FindVariableDeclaratorInNodes(
+            IEnumerable<SyntaxNode> nodes,
+            string variableName)
+        {
+            return nodes
+                .SelectMany(child => child.ChildNodes())
+                .OfType<LocalDeclarationStatementSyntax>()
+                .Select(localDeclaration => localDeclaration.Declaration)
+                .SelectMany(variableDeclaration => variableDeclaration.Variables)
+                .Where(variableDeclarator => variableDeclarator.Identifier.Text.Equals(variableName))
+                .FirstOrDefault();
+        }
+
+        private static IEnumerable<AssignmentExpressionSyntax> FindVariableAssignmentsInNodes(
+            IEnumerable<SyntaxNode> nodes,
+            Func<string, bool> matchVariableName)
+        {
+            return nodes
+                .SelectMany(node => node.ChildNodes())
+                .OfType<ExpressionStatementSyntax>()
+                .Select(expressionStatement => expressionStatement.Expression)
+                .Where(expression => expression.IsKind(SyntaxKind.SimpleAssignmentExpression))
+                .OfType<AssignmentExpressionSyntax>()
+                .Where(expression => matchVariableName(expression.Left.ToString()));
+        }
+
+        private static IEnumerable<AssignmentExpressionSyntax> FindVariableAssignmentsInNodes(
+            IEnumerable<SyntaxNode> nodes,
+            string variableName)
+        {
+            return nodes
+                .SelectMany(node => node.ChildNodes())
+                .OfType<ExpressionStatementSyntax>()
+                .Select(expressionStatement => expressionStatement.Expression)
+                .Where(expression => expression.IsKind(SyntaxKind.SimpleAssignmentExpression))
+                .OfType<AssignmentExpressionSyntax>()
+                .Where(expression => expression.Left.ToString().Equals(variableName));
+        }
+
+        private static SqlStatement ExtractSqlStatement(
+            SyntaxNodeAnalysisContext context,
+            ObjectCreationExpressionSyntax npgsqlCommandCtor)
+        {
+            if (!npgsqlCommandCtor.ArgumentList.Arguments.Any())
+            {
+                /**
+                 * Query is not passed through the constructor,
+                 * therefore it might be assigned to the CommandText prop
+                 */
+                var commandTextAssignment = FindVariableAssignmentsInNodes(
+                    npgsqlCommandCtor.Ancestors(),
+                    name => name.Contains($".{nameof(NpgsqlCommand.CommandText)}"))
+                    .FirstOrDefault();
+                if (commandTextAssignment is null)
+                {
+                    // Query is not assigned to the CommandText prop
+                    return SqlStatement.StatementNotFound;
+                }
+
+                if (commandTextAssignment.Right.IsKind(SyntaxKind.StringLiteralExpression))
+                {
+                    return new SqlStatement(
+                        statement: ExtractQuery(commandTextAssignment.Right.ToString()),
+                        location: commandTextAssignment.Right.GetLocation());
+                }
+                else if (commandTextAssignment.Right.IsKind(SyntaxKind.IdentifierName))
+                {
+                    var varDeclarator = FindVariableDeclaratorInNodes(
+                        nodes: npgsqlCommandCtor.Ancestors(),
+                        variableName: commandTextAssignment.Right.ToString());
+                    var varAssignments = FindVariableAssignmentsInNodes(
+                        nodes: npgsqlCommandCtor.Ancestors(),
+                        variableName: commandTextAssignment.Right.ToString());
+
+                    if (varAssignments.Any())
+                    {
+                        var declarationSymbol = context.SemanticModel.GetDeclaredSymbol(varDeclarator);
+                        int declarationLine = declarationSymbol.Locations.First().GetLineSpan().StartLinePosition.Line;
+                        int commandTextLine = commandTextAssignment.GetLocation().GetLineSpan().StartLinePosition.Line;
+                        var variableAssignment = varAssignments
+                            .OrderBy(assignment =>
+                            {
+                                int assignmentLine = assignment.GetLocation().GetLineSpan().StartLinePosition.Line;
+                                return Math.Abs(commandTextLine - assignmentLine);
+                            })
+                            .First();
+
+                        int assignmentLine = variableAssignment.GetLocation().GetLineSpan().StartLinePosition.Line;
+                        if (Math.Abs(commandTextLine - assignmentLine) < Math.Abs(commandTextLine - declarationLine))
+                        {
+                            return new SqlStatement(
+                                statement: ExtractQuery(variableAssignment.Right.ToString()),
+                                location: variableAssignment.Right.GetLocation());
+                        }
+                    }
+
+                    // The syntax used to assign a value to NpgsqlCommand.CommandText is not supported
+                    return new SqlStatement(
+                        statement: ExtractQuery(varDeclarator.Initializer.Value.ToString()),
+                        location: varDeclarator.Initializer.Value.GetLocation());
+                }
+
+                return default;
+            }
+
+            var queryArgument = npgsqlCommandCtor.ArgumentList.Arguments.First();
+            if (queryArgument.Expression.IsKind(SyntaxKind.StringLiteralExpression))
+            {
+                // Query is defined in the constructor
+                return new SqlStatement(
+                    statement: ExtractQuery(queryArgument.ToString()),
+                    location: npgsqlCommandCtor.GetLocation());
+            }
+
+            /**
+             * The first constructor argument is not a string literal containing the query,
+             * therefore we should search for a variable declaring the query
+             */
+            string queryVariableName = queryArgument.ToString();
+
+            /**
+             * Variable declaration => string query = "";
+             * Variable assignment  => query = "assignment after being declared";
+             */
+            var variableDeclarator = FindVariableDeclaratorInNodes(context.Node.Ancestors(), queryVariableName);
+            var variableAssignments = FindVariableAssignmentsInNodes(context.Node.Ancestors(), queryVariableName);
+
+            if (variableAssignments.Any())
+            {
+                /**
+                 * If there are any assignments that means the variable is reused, we should analyze the one that
+                 * is closest to the NpgsqlCommand constructor
+                 */
+                var declarationSymbol = context.SemanticModel.GetDeclaredSymbol(variableDeclarator);
+                int declarationLine = declarationSymbol.Locations.First().GetLineSpan().StartLinePosition.Line;
+                int npgsqlCommandLine = npgsqlCommandCtor.GetLocation().GetLineSpan().StartLinePosition.Line;
+                var variableAssignment = variableAssignments
+                    .OrderBy(assignment =>
+                    {
+                        int assignmentLine = assignment.GetLocation().GetLineSpan().StartLinePosition.Line;
+                        return Math.Abs(npgsqlCommandLine - assignmentLine);
+                    })
+                    .First();
+
+                int assignmentLine = variableAssignment.GetLocation().GetLineSpan().StartLinePosition.Line;
+                if (Math.Abs(npgsqlCommandLine - assignmentLine) < Math.Abs(npgsqlCommandLine - declarationLine))
+                {
+                    return new SqlStatement(
+                        statement: ExtractQuery(variableAssignment.Right.ToString()),
+                        location: variableAssignment.Right.GetLocation());
+                }
+            }
+
+            return new SqlStatement(
+                statement: ExtractQuery(variableDeclarator.Initializer.Value.ToString()),
+                location: variableDeclarator.Initializer.Value.GetLocation());
+        }
 
         private void AnalyzeInvocationExpressionNode(SyntaxNodeAnalysisContext context)
         {
@@ -84,86 +267,25 @@ namespace NpgsqlAnalyzers
             // Check if object creation is NpgsqlCommand
             if (!(semanticModel.GetSymbolInfo(npgsqlCommandExpression).Symbol is IMethodSymbol methodSymbol) ||
                 methodSymbol.MethodKind != MethodKind.Constructor ||
-                !methodSymbol.ReceiverType.Name.Equals("NpgsqlCommand", StringComparison.OrdinalIgnoreCase))
+                !methodSymbol.ReceiverType.Name.Equals(nameof(NpgsqlCommand), StringComparison.OrdinalIgnoreCase))
             {
                 return;
             }
 
-            // Get first arg for constructor, which is the query
-            var queryArgument = npgsqlCommandExpression.ArgumentList.Arguments.First();
-            if (queryArgument.Expression.IsKind(SyntaxKind.StringLiteralExpression))
+            var statement = ExtractSqlStatement(context, npgsqlCommandExpression);
+            if (statement.IsValid)
             {
-                AnalyzeInConstructorQueryDeclaration(context, queryArgument, npgsqlCommandExpression);
+                ExecuteAndValidateQuery(
+                    query: ReplaceNamedParameters(statement.Statement),
+                    context: context,
+                    sourceLocation: statement.Location);
             }
-            else
+            else if (statement.NotFound)
             {
-                AnalyzeVarQueryDeclaration(context, semanticModel, queryArgument.ToString(), npgsqlCommandExpression);
+                context.ReportDiagnostic(Diagnostic.Create(
+                    descriptor: Rules.MissingCommand,
+                    location: npgsqlCommandExpression.GetLocation()));
             }
-        }
-
-        private void AnalyzeInConstructorQueryDeclaration(
-            SyntaxNodeAnalysisContext context,
-            ArgumentSyntax queryArgumentSyntax,
-            ObjectCreationExpressionSyntax npgsqlCommandExpression)
-        {
-            ExecuteAndValidateQuery(
-                query: ExtractQuery(queryArgumentSyntax.ToString()),
-                context: context,
-                sourceLocation: npgsqlCommandExpression.GetLocation());
-        }
-
-        private void AnalyzeVarQueryDeclaration(
-            SyntaxNodeAnalysisContext context,
-            SemanticModel semanticModel,
-            string variableName,
-            ObjectCreationExpressionSyntax npgsqlCommandExpression)
-        {
-            var variableDeclarator = context.Node
-                .Ancestors()
-                .SelectMany(a => a.ChildNodes())
-                .OfType<LocalDeclarationStatementSyntax>()
-                .Select(localDeclaration => localDeclaration.Declaration)
-                .Select(declaration => declaration.Variables.FirstOrDefault())
-                .OfType<VariableDeclaratorSyntax>()
-                .Where(declarator => declarator.Identifier.Text.Equals(variableName, StringComparison.OrdinalIgnoreCase))
-                .FirstOrDefault();
-            var declarationSymbol = semanticModel.GetDeclaredSymbol(variableDeclarator);
-
-            var variableAssignment = context.Node
-                .Ancestors()
-                .SelectMany(a => a.ChildNodes())
-                .OfType<ExpressionStatementSyntax>()
-                .Where(e => e.Expression.Kind() == SyntaxKind.SimpleAssignmentExpression)
-                .Select(e => e.Expression)
-                .OfType<AssignmentExpressionSyntax>()
-                .Where(e => e.Left.ToString().Equals(variableName, StringComparison.OrdinalIgnoreCase))
-                .FirstOrDefault();
-
-            if (variableAssignment is { })
-            {
-                int declarationLine = declarationSymbol.Locations.First().GetLineSpan().StartLinePosition.Line;
-                int assignmentLine = variableAssignment.GetLocation().GetLineSpan().StartLinePosition.Line;
-                int npgsqlCommandLine = npgsqlCommandExpression.GetLocation().GetLineSpan().StartLinePosition.Line;
-
-                if (Math.Abs(npgsqlCommandLine - assignmentLine) < Math.Abs(npgsqlCommandLine - declarationLine))
-                {
-                    // If re-assignment is closer to where the variable is used, analyze the re-assignment
-                    ExecuteAndValidateQuery(
-                        query: ExtractQuery(variableAssignment.Right.ToString()),
-                        context: context,
-                        sourceLocation: variableAssignment.Right.GetLocation());
-                    return;
-                }
-            }
-
-            ExecuteAndValidateQuery(
-                query: ExtractQuery(variableDeclarator.Initializer.Value.ToString()),
-                context: context,
-                sourceLocation: variableDeclarator.Initializer
-                    .ChildNodes()
-                    .OfType<LiteralExpressionSyntax>()
-                    .First()
-                    .GetLocation());
         }
 
         private void ExecuteAndValidateQuery(
@@ -173,7 +295,7 @@ namespace NpgsqlAnalyzers
         {
             try
             {
-                using var connection = new NpgsqlConnection(_connectionString);
+                using var connection = new NpgsqlConnection(_configuration.ConnectionString);
                 connection.Open();
                 using var command = new NpgsqlCommand(query, connection);
                 command.ExecuteReader(CommandBehavior.SchemaOnly);
@@ -205,10 +327,6 @@ namespace NpgsqlAnalyzers
                             ex.Message));
                         break;
                 }
-            }
-            catch (Exception)
-            {
-                // Ignore
             }
         }
     }